# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Penjadwalan Driver Service API",
        "description": "Service B: API untuk mengelola penjadwalan driver operasional. Bagian dari ekosistem Pencatatan Operasional (Pengisian BBM) - Group 7.",
        "contact": {
            "name": "Hafizh Rafi Maulana Suyufi",
            "email": "102022400210@student.telkomuniversity.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get all driver schedules",
                "description": "Mengambil daftar seluruh jadwal operasional driver.",
                "operationId": "getSchedules",
                "responses": {
                    "200": {
                        "description": "Data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Schedule"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Penjadwalan-Driver-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API Key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized. Invalid or missing API Key (X-IAE-KEY)."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Create a new schedule",
                "description": "Menambah data penugasan atau jadwal baru untuk driver.",
                "operationId": "createSchedule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "driver_name",
                                    "vehicle_id",
                                    "plate_number",
                                    "schedule_date",
                                    "shift"
                                ],
                                "properties": {
                                    "driver_name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "plate_number": {
                                        "type": "string",
                                        "example": "B 1234 ABC"
                                    },
                                    "schedule_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-15"
                                    },
                                    "shift": {
                                        "type": "string",
                                        "example": "pagi",
                                        "enum": [
                                            "pagi",
                                            "siang",
                                            "malam"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active",
                                        "enum": [
                                            "active",
                                            "completed",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Rute Jakarta - Bandung",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Schedule"
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Penjadwalan-Driver-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API Key"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        },
        "/api/v1/schedules/{id}": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get a specific schedule",
                "description": "Mengambil data spesifik jadwal shift berdasarkan ID.",
                "operationId": "getScheduleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Schedule"
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Penjadwalan-Driver-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule not found"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API Key"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Schedule": {
                "title": "Schedule",
                "description": "Driver Schedule model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "driver_name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plate_number": {
                        "type": "string",
                        "example": "B 1234 ABC"
                    },
                    "schedule_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-15"
                    },
                    "shift": {
                        "type": "string",
                        "example": "pagi",
                        "enum": [
                            "pagi",
                            "siang",
                            "malam"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Rute Jakarta - Bandung",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-15T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-15T10:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-IAE-KEY": {
                "type": "apiKey",
                "description": "API Key menggunakan NIM Mahasiswa (102022400210)",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Schedules",
            "description": "Schedules"
        }
    ]
}